+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add commands to jump to the `.yourbase.yml` file for the workspace or
+  active file.
+
 [Unreleased]: https://github.com/yourbase/yourbase-vscode/compare/v0.1.0...HEAD
 
 ## [0.1.0][]